Refactor: Sidebar에서 FiedlSerchBar로 rename
